requirements.txt
flask==2.0.1
google-cloud-datastore
google-cloud-storage
gunicorn
google-cloud-core